requirements.txt
pandas>=2.0.0
numpy==1.26.3
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
plotly>=5.17.0
jupyter>=1.0.0
nbformat>=5.9.0
streamlit==1.31.0
groq==0.4.2